requirements.txt
requests
python-dotenv
pandas
numpy
matplotlib
telepot
schedule
